+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - CHANGELOG added
+- The color of the shape's stroke can now be specified with either one color or an array of colors
 
 ## [1.0.0] - 2020-06-17